Add files via upload
We study electromagnetically induced transparency in Rubidium 87 beyond the three level model considering off-resonant interactions of the control and probe fields. We analyze the D2 line by taking into account optical pumping to the level F=1,m=1, and after when all the Zeeman sublevels are populated. We obtain a curve of storage efficiency of a Gaussian pulse for quantum memory. We do the same with the D1 line.