Add files via upload
In this script, we study electromagnetically induced transparency (EIT) in atomic ensembles of cold atoms of Rubidium 87 on the D2 line. First, we look at the simplest case of the three-level system. Then we consider the case where the atoms are pumped to the Zeeman sublevel F=1,m = 1, and the control field has off-resonant interaction with the excited level F'=3,m=2. Then obtain the storage-retrieval efficiency of a Gaussian light pulse through the ensemble.